requirements.txt
blinker==1.9.0
boto3==1.36.4
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-Cors==5.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
MarkupSafe==3.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.11.1
serpapi==0.1.5
six==1.17.0
urllib3>=1.25.4,<1.27
Werkzeug==3.1.3
